pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.1",
    "langchain>=1.0.8",
    "langchain-google-genai>=3.2.0",
    "langchain-mcp-adapters>=0.1.13",
    "langchain-openai>=1.0.3",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
